cleanup(core): refactor plugin shutdown
